requirements.txt
python_pptx==0.6.21
tqdm==4.64.1
